requirements.txt
Flask
Flask-Dance
python-dotenv
Flask-Mail
geopy
requests
psycopg2-binary
dotenv
google-generativeai